+++ types/domain/entities/KevReportWatermark.d.ts
@@ -1,0 +1,5 @@
+export type KevReportWatermark = {
+    lastReportedDateAdded: string;
+    reportedCveIdsOnLastReportedDateAdded: string[];
+};
+//# sourceMappingURL=KevReportWatermark.d.ts.map