+++ types/domain/usecases/DailySecurityScanUseCase.d.ts
@@ -1,6 +1,7 @@
 import { LocalCommandRunner } from './adapter-interfaces/LocalCommandRunner';
 import { IssueRepository } from './adapter-interfaces/IssueRepository';
 import { HttpRepository } from './adapter-interfaces/HttpRepository';
+import { KevReportWatermarkRepository } from './adapter-interfaces/KevReportWatermarkRepository';
 import { Member } from '../entities/Member';
 export type DailySecurityScanConfig = {
     scanBaseDirectory: string;
@@ -12,7 +13,8 @@ export declare class DailySecurityScanUseCase {
     readonly localCommandRunner: LocalCommandRunner;
     readonly issueRepository: Pick<IssueRepository, 'createNewIssue' | 'searchIssue' | 'createCommentByUrl'>;
     readonly httpRepository: HttpRepository;
-    constructor(localCommandRunner: LocalCommandRunner, issueRepository: Pick<IssueRepository, 'createNewIssue' | 'searchIssue' | 'createCommentByUrl'>, httpRepository: HttpRepository);
+    readonly kevReportWatermarkRepository: KevReportWatermarkRepository;
+    constructor(localCommandRunner: LocalCommandRunner, issueRepository: Pick<IssueRepository, 'createNewIssue' | 'searchIssue' | 'createCommentByUrl'>, httpRepository: HttpRepository, kevReportWatermarkRepository: KevReportWatermarkRepository);
     run: (input: {
         targetDates: Date[];
         org: string;