+++ types/domain/usecases/adapter-interfaces/SilentSessionMessageComposer.d.ts
@@ -7,5 +7,6 @@ export interface SilentSessionMessageComposer {
     composeMainStalledSection: (mainSilentSeconds: number) => string;
     composeMainStalledWithStaleOwnerCallSection: (mainSilentSeconds: number, unansweredOwnerCallAgeSeconds: number) => string;
     composeSubAgentSection: (sections: SubAgentStallSections) => string;
+    composeSubAgentUnconsumedResultSection: (unconsumedResultSubAgents: SubAgentActivity[]) => string;
 }
 //# sourceMappingURL=SilentSessionMessageComposer.d.ts.map